Moved date time pickers and widgets to the same package
