+++ collect_app/src/main/java/org/odk/collect/android/widgets/datetime/DatePickerDetails.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.odk.collect.android.logic;
+package org.odk.collect.android.widgets.datetime;
 
 import java.io.Serializable;
