+++ collect_app/src/main/java/org/odk/collect/android/widgets/datetime/pickers/CustomTimePickerDialog.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.fragments.dialogs;
+package org.odk.collect.android.widgets.datetime.pickers;
 
 import android.app.Dialog;
 import android.app.TimePickerDialog;
