+++ collect_app/src/main/java/org/odk/collect/android/widgets/datetime/pickers/MyanmarDatePickerDialog.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package org.odk.collect.android.fragments.dialogs;
+package org.odk.collect.android.widgets.datetime.pickers;
 
 import org.joda.time.LocalDateTime;
-import org.odk.collect.android.utilities.DateTimeUtils;
+import org.odk.collect.android.widgets.datetime.DateTimeUtils;
 import org.odk.collect.android.utilities.MyanmarDateUtils;
 
 import java.util.List;
