+++ collect_app/src/main/java/org/odk/collect/android/widgets/utilities/DateTimeWidgetUtils.java
@@ -14,16 +14,16 @@ import org.joda.time.chrono.EthiopicChronology;
 import org.joda.time.chrono.IslamicChronology;
 import org.joda.time.chrono.PersianChronologyKhayyamBorkowski;
 import org.odk.collect.android.R;
-import org.odk.collect.android.fragments.dialogs.BikramSambatDatePickerDialog;
-import org.odk.collect.android.fragments.dialogs.BuddhistDatePickerDialog;
-import org.odk.collect.android.fragments.dialogs.CopticDatePickerDialog;
-import org.odk.collect.android.fragments.dialogs.CustomTimePickerDialog;
-import org.odk.collect.android.fragments.dialogs.EthiopianDatePickerDialog;
-import org.odk.collect.android.fragments.dialogs.FixedDatePickerDialog;
-import org.odk.collect.android.fragments.dialogs.IslamicDatePickerDialog;
-import org.odk.collect.android.fragments.dialogs.MyanmarDatePickerDialog;
-import org.odk.collect.android.fragments.dialogs.PersianDatePickerDialog;
-import org.odk.collect.android.logic.DatePickerDetails;
+import org.odk.collect.android.widgets.datetime.pickers.BikramSambatDatePickerDialog;
+import org.odk.collect.android.widgets.datetime.pickers.BuddhistDatePickerDialog;
+import org.odk.collect.android.widgets.datetime.pickers.CopticDatePickerDialog;
+import org.odk.collect.android.widgets.datetime.pickers.CustomTimePickerDialog;
+import org.odk.collect.android.widgets.datetime.pickers.EthiopianDatePickerDialog;
+import org.odk.collect.android.widgets.datetime.pickers.FixedDatePickerDialog;
+import org.odk.collect.android.widgets.datetime.pickers.IslamicDatePickerDialog;
+import org.odk.collect.android.widgets.datetime.pickers.MyanmarDatePickerDialog;
+import org.odk.collect.android.widgets.datetime.pickers.PersianDatePickerDialog;
+import org.odk.collect.android.widgets.datetime.DatePickerDetails;
 import org.odk.collect.android.utilities.Appearances;
 import org.odk.collect.android.utilities.MyanmarDateUtils;
 import org.odk.collect.android.utilities.ThemeUtils;
